moved new artifacts to package
