+++ CurrencyConverter/Transaction.java
@@ -1,11 +1,22 @@
 package CurrencyConverter;
 
-public class Transaction implements TransactionInterface{
+public class Transaction {
 
-    public Transaction(InitialCurrency initialCurrency, InitialCurrency targetCurrency, double amount,double result){}
+    private double amount;
+    private String fromCurrency;
+    private double convertedAmount;
+    private String toCurrency;
+
+    public Transaction(double amount, String fromCurrency, double convertedAmount, String toCurrency) {
+        this.amount = amount;
+        this.fromCurrency = fromCurrency;
+        this.convertedAmount = convertedAmount;
+        this.toCurrency = toCurrency;
+    }
 
     @Override
     public String toString() {
-        return "Transaction{}";
+        return "Обменено: " + amount + " " + fromCurrency + " на " + convertedAmount + " " + toCurrency;
     }
 }
+
